moved neutral functions from vsce to sdk
