+++ packages/sdk/src/index.ts
@@ -13,3 +13,4 @@ export * from "./doc";
 export * from "./AbstractRpcClient";
 export * from "./ZSshClient";
 export * from "./ZSshUtils";
+export * from "./ZClientUtils"
